+++ database/migrations/2025_05_27_082616_create_sale_items_table.php
@@ -18,9 +18,9 @@ return new class extends Migration
             $table->foreignId('sale_id')->constrained();
             $table->foreignId('product_id')->constrained();
             $table->integer('quantity');
-            $table->decimal('sale_price');
-            $table->decimal('discount');
-            $table->decimal('subtotal');
+            $table->double('sale_price');
+            $table->double('discount');
+            $table->double('subtotal');
             $table->foreignId('agency_id')->nullable()->constrained();
             $table->foreignId('created_by')->constrained('users');
             $table->foreignId('user_id');
